refactor: 질문 재생성 리펙토링
- jackson 라이브러리 변경
- 현재 질문이 존재할 때만 보관상태로 변경하도록 분기 처리
- 질문 버저닝 응답 타입 int 에서 long 으로 변경

+++ src/main/java/com/dnd/ahaive/domain/question/repository/QuestionRepository.java
@@ -33,5 +33,5 @@ public interface QuestionRepository extends JpaRepository<Question, Long> {
     int updateQuestionStatusByIdIn(@Param("status") QuestionStatus status, @Param("ids") List<Long> questionIds);
 
     @Query("SELECT MAX(q.version) FROM Question q WHERE q.insight.id = :insightId")
-    Optional<Integer> findMaxVersionByInsightId(@Param("insightId") Long insightId);
+    Optional<Long> findMaxVersionByInsightId(@Param("insightId") Long insightId);
 }
